Create 'models' when passing to cards
