+++ server/libs/graphql-poller.js
@@ -41,7 +41,7 @@ let readyPromise;
 const pollData = (query, name, flags = {}) => {
 	const xFlagsHeader = Object.keys(flags).map(flag => `${flag}:${flags[flag] ? 'on' : 'off'}`).join(',');
 	const poller = new Poller({
-		url: `https://next-graphql-api.ft.com/?apiKey=${process.env.GRAPHQL_API_KEY}`,
+		url: `https://next-graphql-api.ft.com/data?apiKey=${process.env.GRAPHQL_API_KEY}`,
 		options: {
 			method: 'POST',
 			body: JSON.stringify({
