+++ shared/components/card/index.js
@@ -1,0 +1,5 @@
+import Article from './article';
+import Concept from './concept';
+import Video from './video';
+
+export { Article, Concept, Video };
