+++ shared/models/video.js
@@ -1,0 +1,5 @@
+export default item => ({
+	type: 'video',
+	id: item.id,
+	title: item.title
+});
